Abstacted replacement variable editor label, inserter and editor into one component
